ajust some code for golangcli-linter

+++ common/db/go_level_db_test.go
@@ -228,8 +228,8 @@ func BenchmarkRandomReadsWrites(b *testing.B) {
 			idx := (int64(RandInt()) % numItems)
 			internal[idx]++
 			val := internal[idx]
-			idxBytes := int642Bytes(int64(idx))
-			valBytes := int642Bytes(int64(val))
+			idxBytes := int642Bytes(idx)
+			valBytes := int642Bytes(val)
 			//fmt.Printf("Set %X -> %X\n", idxBytes, valBytes)
 			db.Set(
 				idxBytes,
@@ -240,7 +240,7 @@ func BenchmarkRandomReadsWrites(b *testing.B) {
 		{
 			idx := (int64(RandInt()) % numItems)
 			val := internal[idx]
-			idxBytes := int642Bytes(int64(idx))
+			idxBytes := int642Bytes(idx)
 			valBytes, _ := db.Get(idxBytes)
 			//fmt.Printf("Get %X -> %X\n", idxBytes, valBytes)
 			if val == 0 {
